+++ src/main/java/mdteam/ait/registry/ExteriorVariantRegistry.java
@@ -74,6 +74,11 @@ public class ExteriorVariantRegistry extends DatapackRegistry<ExteriorVariantSch
                 buf.encodeAsJson(DatapackVariant.CODEC, variant);
                 continue;
             }
+            if (schema.category() == null) {
+                AITMod.LOGGER.error("Exterior variant " + schema.id() + " has null category!");
+                AITMod.LOGGER.error("Temporarily returning, fix this code!!!"); // todo
+                continue;
+            }
             buf.encodeAsJson(DatapackVariant.CODEC, new DatapackVariant(schema.id(), schema.category().id(), schema.id(), DatapackVariant.DEFAULT_TEXTURE, DatapackVariant.DEFAULT_TEXTURE, false));
         }
         ServerPlayNetworking.send(player, SYNC_TO_CLIENT, buf);
@@ -157,6 +162,8 @@ public class ExteriorVariantRegistry extends DatapackRegistry<ExteriorVariantSch
     public static ExteriorVariantSchema PLINTH_FIRE;
 
     private static void registerDefaults() {
+        // todo make this not static
+
         // TARDIM
         TARDIM_DEFAULT = registerStatic(new TardimDefaultVariant());
         TARDIM_FIRE = registerStatic(new TardimFireVariant());