+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # 1.2.32
 - TARDIS no longer shakes during flight if autopilot is enabled
 - Fixed telepathic circuit ignoring security
+- Fixed issues with monitor exterior selection